apps/conf_entrega/requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Werkzeug==3.0.1
python-dotenv==1.0.1
PyMySQL==1.1.0
cryptography==42.0.5
ftplib==0.0.1